Move JSON5 helper code + deprecate YAML, TOML
- move JSON5 parser to helperFunctions.js
- comment out JSON, TOML, YAML parsers (as discussed 2? months ago
