+++ nmn.Templates/src/configuration.js
@@ -1,9 +1,16 @@
 // @flow
 
-import toml from 'toml'
-import json5 from 'json5'
-import { load } from 'js-yaml'
-import { showMessage, chooseOption } from '../../nmn.sweep/src/userInput'
+import {
+  showMessage,
+  chooseOption,
+  showMessageYesNo
+} from '../../helperFunctions/userInput'
+import {
+  // parseJSON,
+  parseJSON5,
+  // parseTOML,
+  // parseYAML
+} from '../../helperFunctions'
 import {
   getOrMakeTemplateFolder,
   createDefaultConfigNote,
@@ -67,54 +74,49 @@ export async function parseFirstCodeblock(
   // console.log(`\tparseFirstCodeblock: will parse ${contents.length} bytes of ${format}`)
 
   switch (format) {
-    case 'json':
-      return parseJSON(contents)
+    // case 'json':
+    //   return parseJSON(contents)
     case 'json5':
       return parseJSON5(contents)
-    case 'yaml':
-      return parseYAML(contents)
-    case 'toml':
-      return parseTOML(contents)
+    // case 'yaml':
+    //   return parseYAML(contents)
+    // case 'toml':
+    //   return parseTOML(contents)
     default:
-      console.log(
-        `\tparseFirstCodeblock: error: can't deal with format ${format}`,
-      )
+      console.log(`\tparseFirstCodeblock: error: can't deal with format ${format}`)
   }
 }
 
 /**
- * Get configuration section, or if not present, save into _configuration file.
+ * Get configuration section, validating its config if requested.
+ * If configuration section not present, add a default one into the _configuration file (if given)
  * Only deals with json5 case.
  * @author @nmn, @jgclark, @dwertheimer
  * @param {string} configSectionName - name of configuration section to retrieve
- * @param {string} configSectionDefault - JSON5 string to use as default values for this configuration section
- * @param {string} minimumRequiredConfig - contains fields which must exist and type, e.g. "{ openWeatherAPIKey: 'string' }"
- * @return {mixed} return this as structured data, in the format specified by the first line of the first codeblock
+ * @param {string?} configSectionDefault - optional JSON5 string to use as default values for this configuration section
+ * @param {string?} minimumRequiredConfig - optional map of fields which must exist and type, e.g. "{ openWeatherAPIKey: 'string' }"
+ * @return {mixed} return config as structured data, in the format specified by the first line of the first codeblock
  */
 export async function getOrMakeConfigurationSection(
   configSectionName: string,
-  configSectionDefault: string,
+  configSectionDefault: string = '',
   minimumRequiredConfig: { [string]: ?mixed } = {},
 ): Promise<?{ [string]: ?mixed }> {
   let templateFolder = await getOrMakeTemplateFolder()
   if (templateFolder == null) {
-    console.log(
-      `  getOrMakeConfigurationSection: couldn't find the templateFolder ... will try to create it ...`,
-    )
+    console.log(`  getOrMakeConfigurationSection: couldn't find the templateFolder ... will try to create it ...`)
     templateFolder = getOrMakeTemplateFolder()
     return {}
   }
 
-  console.log(`  getOrMakeConfigurationSection: got folder ${templateFolder}`)
+  // console.log(`  getOrMakeConfigurationSection: got folder ${templateFolder}`)
   let configFile = DataStore.projectNotes
     // $FlowIgnore[incompatible-call]
     .filter((n) => n.filename?.startsWith(templateFolder))
     .find((n) => !!n.title?.startsWith('_configuration'))
 
   if (configFile == null) {
-    console.log(
-      `  getOrMakeConfigurationSection: Error: cannot find '_configuration' fil. Will create from default.`,
-    )
+    console.log(`  getOrMakeConfigurationSection: Error: cannot find '_configuration' file. Will create from default.`)
     createDefaultConfigNote()
     configFile = DataStore.projectNotes
       // $FlowIgnore[incompatible-call]
@@ -124,16 +126,14 @@ export async function getOrMakeConfigurationSection(
 
   const content: ?string = configFile?.content
   if (configFile == null || content == null) {
-    console.log(
-      `  getOrMakeConfigurationSection: Error: '_configuration' file not found or empty`,
-    )
+    // Really strange to get here: won't code a response, but will just error.
+    console.log(`  getOrMakeConfigurationSection: Error: '_configuration' file not found or empty`)
     await showMessage(
       `Error: missing or empty '_configuration' file. Please check.`,
     )
-    // Really strange to get here: won't code a response, but will just stop.
     return {}
   }
-  console.log('  getOrMakeConfigurationSection: got _configuration file')
+  // console.log('  getOrMakeConfigurationSection: got _configuration file')
 
   // Get config contents
   const firstCodeblock = content.split('\n```')[1]
@@ -141,13 +141,20 @@ export async function getOrMakeConfigurationSection(
     (await parseFirstCodeblock(firstCodeblock)) ?? {}
 
   // Does it contain the section we want?
-  if (
-    firstCodeblock == null ||
-    config[configSectionName] == // alternative to dot notation that allows variables
-      null
-  ) {
-    // No, so offer to make it and populate it
-    const shouldAddDefaultConfig = await chooseOption<boolean, boolean>(
+  // (use an alternative to dot notation that allows variables)
+  if (firstCodeblock == null || config[configSectionName] == null) {
+    console.log(`  getOrMakeConfigurationSection: no '${configSectionName}' config section found`)
+    // The section is missing.
+    // If no default configuration given, return nothing
+    if (configSectionDefault === '') {
+      console.log(`  getOrMakeConfigurationSection: no default given`)
+      return {}
+    }
+    console.log(`  getOrMakeConfigurationSection: default available`)
+
+    // If a default configuration given, offer to make it and populate it
+    // FIXME: doesn't fire this off
+    const shouldAddDefaultConfig = await chooseOption(
       `No '${configSectionName}' configuration section found.`,
       [
         {
@@ -159,9 +166,13 @@ export async function getOrMakeConfigurationSection(
           value: false,
         },
       ],
-      false,
+      false
     )
-    if (!shouldAddDefaultConfig) {
+    // FIXME: doesn't fire this off
+    // const res = await showMessageYesNo(`Create ${configSectionName} configuration from its defaults?`)
+    console.log(`  getOrMakeConfigurationSection: after SMYN`)
+    // if (res === "No") {
+    if (shouldAddDefaultConfig) {
       return {}
     }
 
@@ -212,8 +223,8 @@ export async function getOrMakeConfigurationSection(
     }
   }
 
-  // We have the configuration, so return it
-  if (Object.keys(minimumRequiredConfig) && config[configSectionName]) {
+  // Yes, we have the configuration, so return it
+  if (Object.keys(minimumRequiredConfig) && config[configSectionName]) { // TODO: ask @dwertheimer about me
     return validateMinimumConfig(
       // $FlowIgnore[incompatible-call]
       config[configSectionName],
@@ -226,8 +237,7 @@ export async function getOrMakeConfigurationSection(
 }
 
 /**
- * Get configuration section, or if not present, save into _configuration file.
- * Only deals with json5 case.
+ * Check whether this config meets a minimum defined spec of keys and types.
  * @author @dwertheimer
  * @param {mixed} config - configuration as structured JSON5 object
  * @param {mixed} validations - JSON5 string to use as default values for this configuration section
@@ -260,67 +270,5 @@ function validateMinimumConfig(
       `    validateMinimumConfig: passed minimum validation spec` // ; config=\n${JSON.stringify(config)}`
     )
     return config
-  }
-}
-
-async function parseJSON(contents: string): Promise<?{ [string]: ?mixed }> {
-  try {
-    return JSON.parse(contents)
-  } catch (e) {
-    console.log(e)
-    await showMessage(
-      'Invalid JSON in your configuration. Please fix it to use configuration',
-    )
-    return {}
-  }
-}
-
-export async function parseJSON5(
-  contents: string,
-): Promise<?{ [string]: ?mixed }> {
-  try {
-    const value = json5.parse(contents)
-    return (value: any)
-  } catch (e) {
-    console.log(e)
-    await showMessage(
-      'Invalid JSON5 in your configuration. Please fix it to use configuration',
-    )
-    return {}
-  }
-}
-
-async function parseYAML(contents: string): Promise<?{ [string]: ?mixed }> {
-  try {
-    const value = load(contents)
-    if (typeof value === 'object') {
-      return (value: any)
-    } else {
-      return {}
-    }
-  } catch (e) {
-    console.log(contents)
-    console.log(e)
-    await showMessage(
-      'Invalid YAML in your configuration. Please fix it to use configuration',
-    )
-    return {}
-  }
-}
-
-async function parseTOML(contents: string): Promise<?{ [string]: ?mixed }> {
-  try {
-    const value = toml.parse(contents)
-    if (typeof value === 'object') {
-      return (value: any)
-    } else {
-      return {}
-    }
-  } catch (e) {
-    console.log(e)
-    await showMessage(
-      'Invalid TOML in your configuration. Please fix it to use configuration',
-    )
-    return {}
   }
 }
